Made a half-hearted attempt at getting empyrean to build 64-bit clean. Add support for the scale= attribute in cal3d config files Draw the player's position at the top of the GameState Gave text fields a grayish background color by default
git-svn-id: file:///desktop/SourceForge%20Backup/empyrean/svnrepos/trunk@267 fea1f14b-4e95-431d-9825-d582385f3f73

+++ empyrean/src/base/ByteBuffer.cpp
@@ -5,8 +5,8 @@
 
 namespace pyr {
 
-    void ByteBuffer::add(const void* buffer, int size) {
-        int new_size = _size + size;
+    void ByteBuffer::add(const void* buffer, size_t size) {
+        size_t new_size = _size + size;
         
         bool do_realloc = false;
         while (_capacity < new_size) {
@@ -21,7 +21,7 @@ namespace pyr {
         _size += size;
     }
     
-    void ByteBuffer::consumeFront(int amount) {
+    void ByteBuffer::consumeFront(size_t amount) {
         amount = std::min(amount, _size);
         memmove(_buffer, _buffer + amount, _size - amount);
         _size -= amount;
